Ändrade int till double

+++ Inlämningsuppgift1/Program.cs
@@ -14,12 +14,12 @@ namespace Inlämningsuppgift1
 
 
             Console.Write("Enter a number:");
-            int num1 = Convert.ToInt32(Console.ReadLine());
+            double num1 = Convert.ToDouble(Console.ReadLine());
 
 
 
             Console.Write("Enter second number:");
-            int num2 = Convert.ToInt32(Console.ReadLine());
+            double num2 = Convert.ToDouble(Console.ReadLine());
 
             if (op == "+")
             {
@@ -39,7 +39,7 @@ namespace Inlämningsuppgift1
             }
 
 
-
+
